Separate index and editor a little
